refactor eval to prevent circular dependencies
